pref: move `lib.init.init` out.
